Separate the extensions for AppKit and UIKit from Image.swift
